refactor: extract ACP manager runtime handle cache
